requirements.txt
argon2-cffi==21.1.0
cffi==1.15.0
click==7.1.2
flufl.enum==4.1.1
numpy==1.21.4
pandas==1.4.1
pycparser==2.21
pyFFTW==0.12.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqtgraph==0.12.3
pyserial==3.5
PyVISA
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
QtPy==1.11.2
regex==2021.11.10
scipy==1.7.3
typing-extensions==4.0.0
python-dateutil==2.8.2

